@@ -1,10 +1,9 @@
-﻿namespace W3ChampionsIdentificationService.Blizzard
+﻿namespace W3ChampionsIdentificationService.Blizzard;
+
+public class BlizzardUserInfo
 {
-    public class BlizzardUserInfo
-    {
-        public string sub { get; set; }
-        public long id { get; set; }
-        public string battletag { get; set; }
-        public string name => battletag.Split("#")[0];
-    }
+    public string sub { get; set; }
+    public long id { get; set; }
+    public string battletag { get; set; }
+    public string name => battletag.Split("#")[0];
 }
