@@ -1,17 +1,16 @@
 ﻿using System;
 
-namespace W3ChampionsIdentificationService.Config
-{
-    public class AppConfig : IAppConfig
-    {
-        public string MongoConnectionString
-        {
-            get => Environment.GetEnvironmentVariable("MONGO_CONNECTION_STRING") ?? "mongodb://localhost:27017";
-        }
+namespace W3ChampionsIdentificationService.Config;
 
-        public string DatabaseName
-        {
-            get => "W3Champions-Identification-Service";
-        }
+public class AppConfig : IAppConfig
+{
+    public string MongoConnectionString
+    {
+        get => Environment.GetEnvironmentVariable("MONGO_CONNECTION_STRING") ?? "mongodb://localhost:27017";
+    }
+
+    public string DatabaseName
+    {
+        get => "W3Champions-Identification-Service";
     }
 }
