@@ -1,8 +1,7 @@
-﻿namespace W3ChampionsIdentificationService.Config
+﻿namespace W3ChampionsIdentificationService.Config;
+
+public interface IAppConfig
 {
-    public interface IAppConfig
-    {
-        string MongoConnectionString { get; }
-        string DatabaseName { get; }
-    }
+    string MongoConnectionString { get; }
+    string DatabaseName { get; }
 }
