@@ -1,14 +1,11 @@
-﻿using System.Collections.Generic;
-using System.Threading.Tasks;
-using W3ChampionsIdentificationService.Identity;
+﻿using System.Threading.Tasks;
 
-namespace W3ChampionsIdentificationService.Identity.Contracts
+namespace W3ChampionsIdentificationService.Identity.Contracts;
+
+public interface IMicrosoftIdentityRepository
 {
-    public interface IMicrosoftIdentityRepository
-    {
-        public Task CreateIndex();
-        public Task<MicrosoftIdentity> GetIdentity(string id);
-        public Task<MicrosoftIdentity> GetIdentityByBattleTag(string battleTag);
-        public Task LinkBattleTag(string id, string battleTag);
-    }
+    public Task CreateIndex();
+    public Task<MicrosoftIdentity> GetIdentity(string id);
+    public Task<MicrosoftIdentity> GetIdentityByBattleTag(string battleTag);
+    public Task LinkBattleTag(string id, string battleTag);
 }
