@@ -1,13 +1,11 @@
 ﻿using MongoDB.Bson.Serialization.Attributes;
-using System.Collections.Generic;
 using W3ChampionsIdentificationService.DatabaseModels;
 
-namespace W3ChampionsIdentificationService.Identity
+namespace W3ChampionsIdentificationService.Identity;
+
+public class MicrosoftIdentity : IIdentifiable
 {
-    public class MicrosoftIdentity : IIdentifiable
-    {
-        [BsonId]
-        public string Id { get; set; }
-        public string battleTag { get; set; }
-    }
+    [BsonId]
+    public string Id { get; set; }
+    public string battleTag { get; set; }
 }
