@@ -3,38 +3,33 @@ using System.Threading.Tasks;
 using W3ChampionsIdentificationService.Config;
 using W3ChampionsIdentificationService.Identity.Contracts;
 
-namespace W3ChampionsIdentificationService.Identity
+namespace W3ChampionsIdentificationService.Identity.Repositories;
+
+public class MicrosoftIdentityRepository(MongoClient mongoClient, IAppConfig appConfig) : MongoDbRepositoryBase(mongoClient, appConfig), IMicrosoftIdentityRepository
 {
-    public class MicrosoftIdentityRepository : MongoDbRepositoryBase, IMicrosoftIdentityRepository
+    public async Task CreateIndex()
     {
-        public MicrosoftIdentityRepository(MongoClient mongoClient, IAppConfig appConfig) : base(mongoClient, appConfig)
-        {
-        }
+        var mongoCollection = CreateCollection<MicrosoftIdentity>();
+        var indexKeysDefinition = Builders<MicrosoftIdentity>.IndexKeys.Ascending(entity => entity.battleTag);
+        await mongoCollection.Indexes.CreateOneAsync(new CreateIndexModel<MicrosoftIdentity>(indexKeysDefinition));
+    }
 
-        public async Task CreateIndex()
-        {
-            var mongoCollection = CreateCollection<MicrosoftIdentity>();
-            var indexKeysDefinition = Builders<MicrosoftIdentity>.IndexKeys.Ascending(entity => entity.battleTag);
-            await mongoCollection.Indexes.CreateOneAsync(new CreateIndexModel<MicrosoftIdentity>(indexKeysDefinition));
-        }
+    public async Task<MicrosoftIdentity> GetIdentity(string id)
+    {
+        return await LoadFirst<MicrosoftIdentity>(x => x.Id == id);
+    }
 
-        public async Task<MicrosoftIdentity> GetIdentity(string id)
-        {
-            return await LoadFirst<MicrosoftIdentity>(x => x.Id == id);
-        }
+    public async Task<MicrosoftIdentity> GetIdentityByBattleTag(string battleTag)
+    {
+        return await LoadFirst<MicrosoftIdentity>(x => x.battleTag == battleTag);
+    }
 
-        public async Task<MicrosoftIdentity> GetIdentityByBattleTag(string battleTag)
+    public async Task LinkBattleTag(string id, string battleTag)
+    {
+        await Upsert(new MicrosoftIdentity
         {
-            return await LoadFirst<MicrosoftIdentity>(x => x.battleTag == battleTag);
-        }
-
-        public async Task LinkBattleTag(string id, string battleTag)
-        {
-            await Upsert(new MicrosoftIdentity
-            {
-                Id = id,
-                battleTag = battleTag
-            });
-        }
+            Id = id,
+            battleTag = battleTag
+        });
     }
 }
