@@ -1,7 +1,6 @@
-﻿namespace W3ChampionsIdentificationService.Microsoft
+﻿namespace W3ChampionsIdentificationService.Microsoft;
+
+public class MicrosoftUser
 {
-    public class MicrosoftUser
-    {
-        public string sub { get; set; }
-    }
+    public string sub { get; set; }
 }
