@@ -4,24 +4,18 @@ using System;
 using System.Threading.Tasks;
 using System.Threading;
 using W3ChampionsIdentificationService.Identity.Contracts;
-using static System.Formats.Asn1.AsnWriter;
 
-namespace W3ChampionsIdentificationService
+namespace W3ChampionsIdentificationService;
+
+public class MigratorHostedService(IServiceProvider serviceProvider) : IHostedService
 {
-    public class MigratorHostedService : IHostedService
+    private readonly IServiceProvider _serviceProvider = serviceProvider;
+
+    public async Task StartAsync(CancellationToken cancellationToken)
     {
-        private readonly IServiceProvider _serviceProvider;
-        public MigratorHostedService(IServiceProvider serviceProvider)
-        {
-            _serviceProvider = serviceProvider;
-        }
-
-        public async Task StartAsync(CancellationToken cancellationToken)
-        {
-            await _serviceProvider.GetService<IMicrosoftIdentityRepository>().CreateIndex();
-        }
-
-        // noop
-        public Task StopAsync(CancellationToken cancellationToken) => Task.CompletedTask;
+        await _serviceProvider.GetService<IMicrosoftIdentityRepository>().CreateIndex();
     }
+
+    // noop
+    public Task StopAsync(CancellationToken cancellationToken) => Task.CompletedTask;
 }
