@@ -1,56 +1,48 @@
-﻿using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using System.Threading.Tasks;
 using W3ChampionsIdentificationService.Middleware;
 using W3ChampionsIdentificationService.RolesAndPermissions.Contracts;
 
-namespace W3ChampionsIdentificationService.RolesAndPermissions.CommandHandlers
+namespace W3ChampionsIdentificationService.RolesAndPermissions.CommandHandlers;
+
+public class UsersCommandHandler(
+    IUsersRepository usersRepository,
+    RolesAndPermissionsValidator validator) : IUsersCommandHandler
 {
-    public class UsersCommandHandler : IUsersCommandHandler
+    private readonly IUsersRepository _usersRepository = usersRepository;
+    private readonly RolesAndPermissionsValidator _validator = validator;
+
+    public async Task CreateUser(User user)
     {
-        private readonly IUsersRepository _usersRepository;
-        private readonly RolesAndPermissionsValidator _validator;
-        public UsersCommandHandler(
-            IUsersRepository usersRepository,
-            RolesAndPermissionsValidator validator)
+        await _validator.ValidateRoleList(user.Roles);
+        await _validator.ValidateCreateUser(user);
+
+        await _usersRepository.CreateUser(new User()
         {
-            _usersRepository = usersRepository;
-            _validator = validator;
-        }
+            Id = user.Id,
+            Roles = user.Roles,
+        });
+    }
 
-        public async Task CreateUser(User user)
+    public async Task UpdateUser(User user)
+    {
+        await _validator.ValidateRoleList(user.Roles);
+        await _validator.ValidateUpdateUser(user);
+
+        await _usersRepository.UpdateUser(new User()
         {
-            await _validator.ValidateRoleList(user.Roles);
-            await _validator.ValidateCreateUser(user);
+            Id = user.Id,
+            Roles = user.Roles,
+        });
+    }
 
-            await _usersRepository.CreateUser(new User()
-            {
-                Id = user.Id,
-                Roles = user.Roles,
-            });
-        }
-
-        public async Task UpdateUser(User user)
-        {   
-            await _validator.ValidateRoleList(user.Roles);
-            await _validator.ValidateUpdateUser(user);
-
-            await _usersRepository.UpdateUser(new User()
-            {
-                Id = user.Id,
-                Roles = user.Roles,
-            });
-        }
-
-        public async Task DeleteUser(string id)
+    public async Task DeleteUser(string id)
+    {
+        var user = await _usersRepository.GetUser(id);
+        if (user == null)
         {
-            var user = await _usersRepository.GetUser(id);
-            if (user == null)
-            {
-                throw new HttpException(404, $"Role with id: '{id}' not found");
-            }
-
-            await _usersRepository.DeleteUser(id);
+            throw new HttpException(404, $"Role with id: '{id}' not found");
         }
+
+        await _usersRepository.DeleteUser(id);
     }
 }
