@@ -1,11 +1,10 @@
 ﻿using System.Threading.Tasks;
 
-namespace W3ChampionsIdentificationService.RolesAndPermissions.Contracts
+namespace W3ChampionsIdentificationService.RolesAndPermissions.Contracts;
+
+public interface IPermissionsCommandHandler
 {
-    public interface IPermissionsCommandHandler
-    {
-        public Task CreatePermission(Permission role);
-        public Task UpdatePermission(Permission role);
-        public Task DeletePermission(string id);
-    }
+    public Task CreatePermission(Permission role);
+    public Task UpdatePermission(Permission role);
+    public Task DeletePermission(string id);
 }
