@@ -1,41 +1,35 @@
 ﻿using MongoDB.Driver;
-using System.Linq;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using W3ChampionsIdentificationService.RolesAndPermissions.Contracts;
 using W3ChampionsIdentificationService.Config;
 
-namespace W3ChampionsIdentificationService.RolesAndPermissions
+namespace W3ChampionsIdentificationService.RolesAndPermissions.Repositories;
+
+public class UsersRepository(MongoClient mongoClient, IAppConfig appConfig) : MongoDbRepositoryBase(mongoClient, appConfig), IUsersRepository
 {
-    public class UsersRepository : MongoDbRepositoryBase, IUsersRepository
+    public async Task<User> GetUser(string id)
     {
-        public UsersRepository(MongoClient mongoClient, IAppConfig appConfig) : base(mongoClient, appConfig)
-        {
-        }
+        return await LoadFirst<User>(id);
+    }
 
-        public async Task<User> GetUser(string id)
-        {
-            return await LoadFirst<User>(id);
-        }
+    public async Task<List<User>> GetAllUsers(int? limit = 50, int? offset = 0)
+    {
+        return await LoadAll<User>(null, limit, offset);
+    }
 
-        public async Task<List<User>> GetAllUsers(int? limit = 50, int? offset = 0)
-        {
-            return await LoadAll<User>(null, limit, offset);
-        }
+    public async Task CreateUser(User user)
+    {
+        await Insert(user);
+    }
 
-        public async Task CreateUser(User user)
-        {
-            await Insert(user);
-        }
+    public async Task UpdateUser(User user)
+    {
+        await Upsert(user);
+    }
 
-        public async Task UpdateUser(User user)
-        {
-            await Upsert(user);
-        }
-
-        public async Task DeleteUser(string id)
-        {
-            await Delete<User>(id);
-        }
+    public async Task DeleteUser(string id)
+    {
+        await Delete<User>(id);
     }
 }
