@@ -1,9 +1,8 @@
 using System.Threading.Tasks;
 
-namespace W3ChampionsIdentificationService.Twitch
+namespace W3ChampionsIdentificationService.Twitch;
+
+public interface ITwitchAuthenticationService
 {
-    public interface ITwitchAuthenticationService
-    {
-        Task<OAuthToken> GetToken();
-    }
+    Task<OAuthToken> GetToken();
 }
