@@ -1,6 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
@@ -11,114 +9,102 @@ using W3ChampionsIdentificationService.RolesAndPermissions;
 using W3ChampionsIdentificationService.RolesAndPermissions.Contracts;
 using W3ChampionsIdentificationService.Twitch;
 
-namespace W3ChampionsIdentificationService.W3CAuthentication
+namespace W3ChampionsIdentificationService.W3CAuthentication;
+
+[ApiController]
+[Route("api/oauth")]
+public class AuthorizationController(
+    IBlizzardAuthenticationService blizzardAuthenticationService,
+    ITwitchAuthenticationService twitchAuthenticationService,
+    IMicrosoftAuthenticationService microsoftAuthenticationService,
+    IUsersRepository usersRepository,
+    IRolesRepository rolesRepository,
+    IPermissionsRepository permissionsRepository,
+    IMicrosoftIdentityRepository microsoftIdentityRepository) : ControllerBase
 {
+    private readonly IBlizzardAuthenticationService _blizzardAuthenticationService = blizzardAuthenticationService;
+    private readonly ITwitchAuthenticationService _twitchAuthenticationService = twitchAuthenticationService;
+    private readonly IMicrosoftAuthenticationService _microsoftAuthenticationService = microsoftAuthenticationService;
+    private readonly IUsersRepository _usersRepository = usersRepository;
+    private readonly IRolesRepository _rolesRepository = rolesRepository;
+    private readonly IPermissionsRepository _permissionsRepository = permissionsRepository;
+    private readonly IMicrosoftIdentityRepository _microsoftIdentityRepository = microsoftIdentityRepository;
 
-    [ApiController]
-    [Route("api/oauth")]
-    public class AuthorizationController : ControllerBase
+    private static readonly string JwtPrivateKey = Regex.Unescape(Environment.GetEnvironmentVariable("JWT_PRIVATE_KEY") ?? "");
+    private static readonly string JwtPublicKey = Regex.Unescape(Environment.GetEnvironmentVariable("JWT_PUBLIC_KEY") ?? "");
+
+    [HttpGet("token")]
+    public async Task<IActionResult> GetBlizzardToken(
+        [FromQuery] string code,
+        [FromQuery] string redirectUri,
+        [FromQuery] BnetRegion region)
     {
-        private readonly IBlizzardAuthenticationService _blizzardAuthenticationService;
-        private readonly ITwitchAuthenticationService _twitchAuthenticationService;
-        private readonly IMicrosoftAuthenticationService _microsoftAuthenticationService;
-        private readonly IUsersRepository _usersRepository;
-        private readonly IRolesRepository _rolesRepository;
-        private readonly IPermissionsRepository _permissionsRepository;
-        private readonly IMicrosoftIdentityRepository _microsoftIdentityRepository;
-
-        private static readonly string JwtPrivateKey = Regex.Unescape(Environment.GetEnvironmentVariable("JWT_PRIVATE_KEY") ?? "");
-        private static readonly string JwtPublicKey = Regex.Unescape(Environment.GetEnvironmentVariable("JWT_PUBLIC_KEY") ?? "");
-
-        public AuthorizationController(
-            IBlizzardAuthenticationService blizzardAuthenticationService,
-            ITwitchAuthenticationService twitchAuthenticationService,
-            IMicrosoftAuthenticationService microsoftAuthenticationService,
-            IUsersRepository usersRepository,
-            IRolesRepository rolesRepository,
-            IPermissionsRepository permissionsRepository,
-            IMicrosoftIdentityRepository microsoftIdentityRepository)
+        var token = await _blizzardAuthenticationService.GetToken(code, redirectUri, region);
+        if (token == null)
         {
-            _blizzardAuthenticationService = blizzardAuthenticationService;
-            _twitchAuthenticationService = twitchAuthenticationService;
-            _microsoftAuthenticationService = microsoftAuthenticationService;
-            _usersRepository = usersRepository;
-            _rolesRepository = rolesRepository;
-            _permissionsRepository = permissionsRepository;
-            _microsoftIdentityRepository = microsoftIdentityRepository;
+            return Unauthorized("Sorry H4ckerb0i");
         }
 
-        [HttpGet("token")]
-        public async Task<IActionResult> GetBlizzardToken(
-            [FromQuery] string code,
-            [FromQuery] string redirectUri,
-            [FromQuery] BnetRegion region)
+        var userInfo = await _blizzardAuthenticationService.GetUser(token.access_token, region);
+        if (userInfo == null)
         {
-            var token = await _blizzardAuthenticationService.GetToken(code, redirectUri, region);
-            if (token == null)
-            {
-                return Unauthorized("Sorry H4ckerb0i");
-            }
-
-            var userInfo = await _blizzardAuthenticationService.GetUser(token.access_token, region);
-            if (userInfo == null)
-            {
-                return Unauthorized("Sorry H4ckerb0i");
-            }
-
-            var user = await _usersRepository.GetUser(userInfo.battletag);
-            var permissions = await _permissionsRepository.GetPermissionsForAdmin(userInfo.battletag);
-
-            // Save user's Battle.net account id to the database
-            if (string.IsNullOrEmpty(user?.BnetId))
-            {
-                await _usersRepository.UpdateUser(new User { 
-                    Id = userInfo.battletag,
-                    BnetId = userInfo.id.ToString(),
-                });
-            }
-
-            var w3User = W3CUserAuthentication.Create(userInfo.battletag, JwtPrivateKey, permissions, userInfo.id);
-
-            return Ok(w3User);
+            return Unauthorized("Sorry H4ckerb0i");
         }
 
-        [HttpGet("token-microsoft")]
-        public async Task<IActionResult> GetMicrosoftToken(
-            [FromQuery] string code,
-            [FromQuery] string redirectUri)
+        var user = await _usersRepository.GetUser(userInfo.battletag);
+        var permissions = await _permissionsRepository.GetPermissionsForAdmin(userInfo.battletag);
+
+        // Save user's Battle.net account id to the database
+        if (string.IsNullOrEmpty(user?.BnetId))
         {
-            var token = await _microsoftAuthenticationService.GetIdToken(code, redirectUri);
-            if (token == null)
+            await _usersRepository.UpdateUser(new User
             {
-                return Unauthorized("Sorry H4ckerb0i");
-            }
-            var u = await _microsoftAuthenticationService.GetUser(token);
-
-            var userInfo = await _microsoftIdentityRepository.GetIdentity(u.sub);
-            if (userInfo == null)
-            {
-                return Unauthorized("Not Linked");
-            }
-
-            var permissions = await _permissionsRepository.GetPermissionsForAdmin(userInfo.battleTag);
-
-            var w3User = W3CUserAuthentication.Create(userInfo.battleTag, JwtPrivateKey, permissions);
-
-            return Ok(w3User);
+                Id = userInfo.battletag,
+                BnetId = userInfo.id.ToString(),
+            });
         }
 
-        [HttpGet("user-info")]
-        public IActionResult GetUserInfo([FromQuery] string jwt)
+        var w3User = W3CUserAuthentication.Create(userInfo.battletag, JwtPrivateKey, permissions, userInfo.id);
+
+        return Ok(w3User);
+    }
+
+    [HttpGet("token-microsoft")]
+    public async Task<IActionResult> GetMicrosoftToken(
+        [FromQuery] string code,
+        [FromQuery] string redirectUri)
+    {
+        var token = await _microsoftAuthenticationService.GetIdToken(code, redirectUri);
+        if (token == null)
         {
-            var user = W3CUserAuthentication.FromJWT(jwt, JwtPublicKey);
-            return user != null ? (IActionResult) Ok(user) : Unauthorized("Sorry Hackerboi");
+            return Unauthorized("Sorry H4ckerb0i");
+        }
+        var u = await _microsoftAuthenticationService.GetUser(token);
+
+        var userInfo = await _microsoftIdentityRepository.GetIdentity(u.sub);
+        if (userInfo == null)
+        {
+            return Unauthorized("Not Linked");
         }
 
-        [HttpGet("twitch")]
-        public async Task<IActionResult> GetTwitchToken()
-        {
-            var token = await _twitchAuthenticationService.GetToken();
-            return Ok(token);
-        }
+        var permissions = await _permissionsRepository.GetPermissionsForAdmin(userInfo.battleTag);
+
+        var w3User = W3CUserAuthentication.Create(userInfo.battleTag, JwtPrivateKey, permissions);
+
+        return Ok(w3User);
+    }
+
+    [HttpGet("user-info")]
+    public IActionResult GetUserInfo([FromQuery] string jwt)
+    {
+        var user = W3CUserAuthentication.FromJWT(jwt, JwtPublicKey);
+        return user != null ? Ok(user) : Unauthorized("Sorry Hackerboi");
+    }
+
+    [HttpGet("twitch")]
+    public async Task<IActionResult> GetTwitchToken()
+    {
+        var token = await _twitchAuthenticationService.GetToken();
+        return Ok(token);
     }
 }
